+++ src/index.html
@@ -1,44 +1,47 @@
 <!doctype html>
 <html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Sleep Outside | Home</title>
+    <link rel="stylesheet" href="/css/style.css" />
+    <script type="module" src="./js/main.js"></script>
+  </head>
 
-<head>
-  <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>Sleep Outside | Home</title>
-  <link rel="stylesheet" href="/css/style.css" />
-  <script type="module" src="./js/main.js"></script>
-</head>
-
-<body>
-  <header class="divider" id="main-header"><!-- Using templates in /partials/header.html --></header>
-  <main class="divider">
-    <div class="hero">
-      <img src="/images/banner-sm.jpg" srcset="images/banner-sm.jpg 500w, images/banner.jpg"
-        alt="image of a high mountain lake" />
-      <div class="logo logo--square">
-        <img src="/images/noun_Tent_2517.svg" alt="tent image for logo" />
-        <div>Sleep<span class="highlight">Outside</span></div>
+  <body>
+    <header class="divider" id="main-header">
+      <!-- Using templates in /partials/header.html -->
+    </header>
+    <main class="divider">
+      <div class="hero">
+        <img
+          src="/images/banner-sm.jpg"
+          srcset="images/banner-sm.jpg 500w, images/banner.jpg"
+          alt="image of a high mountain lake"
+        />
+        <div class="logo logo--square">
+          <img src="/images/noun_Tent_2517.svg" alt="tent image for logo" />
+          <div>Sleep<span class="highlight">Outside</span></div>
+        </div>
       </div>
-    </div>
-    <section class="mission">
-      <p>We spend too much time these days indoors...sitting usually.</p>
-      <p>No wonder we are a mess!</p>
-      <p>
-        What we really need is to get outside...in fact SLEEP outside. Our
-        mission is to provide you with everything you need to do this
-        comfortably at an affordable price.
-      </p>
-    </section>
-    <section class="products">
-      <h2>Top Products</h2>
-      <ul class="product-list">
-        <!-- Inserted dinamically from ProductList -->
-      </ul>
-    </section>
-  </main>
-  <footer id="main-footer">
-    <!-- Using templates in /partials/footer.html -->
-  </footer>
-</body>
-
+      <section class="mission">
+        <p>We spend too much time these days indoors...sitting usually.</p>
+        <p>No wonder we are a mess!</p>
+        <p>
+          What we really need is to get outside...in fact SLEEP outside. Our
+          mission is to provide you with everything you need to do this
+          comfortably at an affordable price.
+        </p>
+      </section>
+      <section class="products">
+        <h2>Top Products</h2>
+        <ul class="product-list">
+          <!-- Inserted dinamically from ProductList -->
+        </ul>
+      </section>
+    </main>
+    <footer id="main-footer">
+      <!-- Using templates in /partials/footer.html -->
+    </footer>
+  </body>
 </html>
