+++ src/js/checkout.js
@@ -1,0 +1,3 @@
+import { loadHeaderFooter } from "./utils.mjs";
+
+loadHeaderFooter();
