+++ src/js/main.js
@@ -2,9 +2,10 @@ import ProductData from "./ProductData.mjs";
 import ProductList from "./ProductList.mjs";
 import { loadHeaderFooter } from "./utils.mjs";
 
+loadHeaderFooter();
+
 const dataSource = new ProductData("tents");
 const element = document.querySelector(".product-list");
 const listing = new ProductList("Tents", dataSource, element);
 
 listing.init();
-loadHeaderFooter();
